Fix: Cast overall_score to int before passing to score_to_status()
round() returns a float, but score_to_status(int , ...) has a
strict type hint. This caused a fatal crash when rendering the dashboard
after a successful scan.

Found via PHP Fatal error in debug.log:
  'Argument #1 () must be of type int, float given'

+++ includes/class-report-generator.php
@@ -65,7 +65,7 @@ class PHPCC_Report_Generator {
             }
         }
 
-        $overall_score = $total > 0 ? round(array_sum(array_column($results, 'readiness_score')) / $total) : 0;
+        $overall_score = $total > 0 ? (int) round(array_sum(array_column($results, 'readiness_score')) / $total) : 0;
         $overall_status = $this->score_to_status($overall_score, $critical);
 
         return [
